ev: use usize (not u64) for shared socket accounting counters
The u64 atomics in epoll/kqueue SharedState fail to compile on 32-bit
targets (arm/riscv32/thumb): @atomicLoad/@atomicRmw reject 64-bit ints
there. Use usize to match the per-loop LoopState counters - u32 on 32-bit
(native atomic), u64 on 64-bit.

+++ src/ev/backends/kqueue.zig
@@ -50,8 +50,8 @@ pub const SharedState = struct {
     /// Group-shared accounting: a completion submitted on one loop may be
     /// finished by the loop that owns the fd registration, so active/inflight_io
     /// are shared atomics rather than per-LoopState fields (see LoopState).
-    active: std.atomic.Value(u64) = .init(0),
-    inflight_io: std.atomic.Value(u64) = .init(0),
+    active: std.atomic.Value(usize) = .init(0),
+    inflight_io: std.atomic.Value(usize) = .init(0),
     /// Cross-loop single-owner socket registration table, shared by every loop
     /// in the group. See sockreg.zig.
     sock_table: sockreg.Table = .{},
